match sliders
We'd like to add a specific status when the matches are paused in the
match sliders and match listing tables. When the match is paused, the
red dot will not blink and the text will be gray and read "HT". Like you
did for the ".live", can you add a match with a ".paused" class in the
match slider with these modifications. See links for details